+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2022-06-30
+### Added
+- Script that converts metadata files into a delimiter separated table, where:
+  - Each metadata field is a column
+  - Each row is a single metadata file
+
 ## [1.4.1] - 2022-06-27
 ### Fixed
 - The issue for when ORs are 0, and added a more explicit modification of ORU and ORL